методы перенесены в utils, код работает, тестовый тест работает
